requirements.txt
tensorflow-gpu==1.8.0
wget==3.2
pandas==0.23.0
Keras==2.1.6
opencv-python==3.4.1.15
h5py==2.7.1
numpy==1.13.3
scikit-learn==0.19.1
Pillow==3.4.2
